refactor: move bash command execution to the command executor
This required command executor refactor (done)
