+++ src/streetrace/commands/base_command.py
@@ -32,12 +32,14 @@ class Command(ABC):
         """A short description of what the command does."""
 
     @abstractmethod
-    async def execute_async(self) -> None:
+    async def execute_async(self, user_input: str) -> str | None:
         """Execute the command's action.
 
         Args:
-            app_instance: The main Application instance.
-                          Commands can use this to access application state or methods.
+            user_input: The raw input string from the user (e.g., "/exit").
+
+        Returns:
+            None or any output to send to the model.
 
         Raises:
             SystemExit: System exit requested by the command.
