+++ src/streetrace/commands/definitions/bash_command.py
@@ -1,0 +1,131 @@
+"""Implement the bash command to run in a terminal session.
+
+This module defines the BashCommand class which allows users to run a bash command
+in a terminal session.
+"""
+
+import subprocess
+from typing import override
+
+from streetrace.commands.base_command import Command
+from streetrace.log import get_logger
+from streetrace.terminal_session import SessionData, SessionEvent, TerminalSession
+
+logger = get_logger(__name__)
+
+
+class BashCommand(Command):
+    """Command to run a bash command in a virtual terminal.
+
+    Handles special command syntax: any user input starting from ! is treated as a bash
+    command.
+    """
+
+    @property
+    def names(self) -> list[str]:
+        """Command invocation names."""
+        return ["!"]
+
+    @property
+    def description(self) -> str:
+        """Command description."""
+        return "Run a bash command in a virtual terminal."
+
+    @override
+    async def execute_async(self, user_input: str) -> str | None:
+        """Signal the application to stop.
+
+        Args:
+            user_input: The raw input string from the user (e.g., "/exit").
+
+        Raises:
+            SystemExit to signal exit.
+
+        """
+        cli_command = user_input.strip()
+        if cli_command[0] == "!":
+            cli_command = cli_command[1:]
+
+        logger.info(
+            "Executing CLI command: %s",
+            cli_command,
+            extra={"command": cli_command},
+        )
+
+        # Collect command output
+        command_return_code = None
+        command_error = None
+
+        def on_session_complete(event: SessionEvent) -> None:
+            nonlocal command_return_code, command_error
+            command_return_code = event.return_code
+            if event.error_message:
+                command_error = event.error_message
+
+        # Create terminal session
+        terminal_session = TerminalSession(
+            on_session_complete=on_session_complete,
+        )
+
+        try:
+            # Start the terminal session
+            terminal_session.start()
+
+            # Execute the command
+            return_code = terminal_session.execute_command(cli_command)
+
+            # Format the output for the model
+            return self._format_cli_output(
+                cli_command,
+                terminal_session.session_data,
+                return_code,
+                command_error,
+            )
+
+        except (OSError, subprocess.SubprocessError) as e:
+            logger.exception(
+                "CLI command execution failed",
+                extra={"command": cli_command, "error": str(e)},
+            )
+            raise
+        finally:
+            # Clean up terminal session
+            terminal_session.stop()
+
+    def _format_cli_output(
+        self,
+        command: str,
+        session_data: list[SessionData],
+        return_code: int,
+        error_message: str | None = None,
+    ) -> str:
+        """Format CLI command output for the model.
+
+        Args:
+            command: The executed command
+            session_data: Session data from TerminalSession
+            return_code: Command return code
+            error_message: Error message if any
+
+        Returns:
+            Formatted string to send to the model
+
+        """
+        lines = [f"Command: {command}"]
+
+        if return_code is not None:
+            lines.append(f"Exit code: {return_code}")
+
+        if error_message:
+            lines.append(f"Error: {error_message}")
+
+        # Add session data output
+        if session_data:
+            lines.append("Output:")
+            lines.extend(
+                data.content for data in session_data if data.source == "command"
+            )
+        else:
+            lines.append("Output: (no output)")
+
+        return "\n".join(lines)
