+++ src/streetrace/commands/definitions/exit_command.py
@@ -29,11 +29,11 @@ class ExitCommand(Command):
         return "Exit the interactive session."
 
     @override
-    async def execute_async(self) -> None:
+    async def execute_async(self, user_input: str) -> str | None:
         """Signal the application to stop.
 
         Args:
-            _: The application instance (required by the interface but unused).
+            user_input: The raw input string from the user (e.g., "/exit").
 
         Raises:
             SystemExit to signal exit.
