+++ src/streetrace/commands/definitions/reset_command.py
@@ -38,11 +38,11 @@ class ResetSessionCommand(Command):
         )
 
     @override
-    async def execute_async(self) -> None:
+    async def execute_async(self, user_input: str) -> str | None:
         """Execute the history clearing action using the HistoryManager.
 
         Args:
-            app_instance: The main Application instance.
+            user_input: The raw input string from the user (e.g., "/exit").
 
         """
         # Re-initialize history as if starting an interactive session
@@ -53,3 +53,4 @@ class ResetSessionCommand(Command):
                 "Session was reset. Start conversation to create a new session.",
             ),
         )
+        return None
